common_base::bit_vec::prelude

Struct BitPtrRange

#[repr(C)]
pub struct BitPtrRange<M = Const, T = usize, O = Lsb0>
where
    M: Mutability,
    T: BitStore,
    O: BitOrder,
{
    pub start: BitPtr<M, T, O>,
    pub end: BitPtr<M, T, O>,
}

Bit-Pointer Range

This type is equivalent in purpose, but superior in functionality, to Range<BitPtr<M, T, O>>. If the standard library stabilizes Step, the trait used to drive Range operations, then this type will likely be destroyed in favor of an impl Step for BitPtr block and use of standard ranges.

Like Range, this is a half-open set where the low bit-pointer selects the first live bit in a span and the high bit-pointer selects the first dead bit after the span.

This type is not capable of inspecting provenance, and has no requirement of its own that both bit-pointers be derived from the same provenance region. It is safe to construct and use with any pair of bit-pointers; however, the bit-pointers it produces are, necessarily, unsafe to use.

Original

Range<*bool>

Memory Representation

BitPtr is required to be repr(packed) in order to satisfy the BitRef size optimizations. In order to stay minimally sized itself, this type has no alignment requirement, and reading either bit-pointer may incur a misalignment penalty. Reads are always safe and valid; they may merely be slow.

Type Parameters

This takes the same type parameters as BitPtr, as it is simply a pair of bit-pointers with range semantics.

Fields
start: BitPtr<M, T, O>

The lower, inclusive, bound of the range. The bit to which this points is considered live.

end: BitPtr<M, T, O>

The higher, exclusive, bound of the range. The bit to which this points is considered dead, and the pointer may be one bit beyond the bounds of an allocation region.

Because Rust and LLVM both define the address of base + (len * width) as being within the provenance of base, even though that address may itself be the base address of another region in a different provenance, and bit-pointers are always composed of an ordinary memory address and a bit-counter, the ending bit-pointer is always valid.

Implementations

impl<M, T, O> BitPtrRange<M, T, O>
where M: Mutability, T: BitStore, O: BitOrder,

pub const EMPTY: BitPtrRange<M, T, O>

The canonical empty range. All ranges with zero length (equal .start and .end) are equally empty.

pub fn from_range(_: Range<BitPtr<M, T, O>>) -> BitPtrRange<M, T, O>

Explicitly converts a Range<BitPtr> into a BitPtrRange.

pub fn into_range(self) -> Range<BitPtr<M, T, O>>

Explicitly converts a BitPtrRange into a Range<BitPtr>.

pub fn is_empty(&self) -> bool

Tests if the range is empty (the distance between bit-pointers is 0).

Original

Range::is_empty

Examples

use bitvec::prelude::*;
use bitvec::ptr::BitPtrRange;

let data = 0u8;
let bp = BitPtr::<_, _, Lsb0>::from_ref(&data);
let mut range = BitPtrRange::from_range(bp .. bp.wrapping_add(1));

assert!(!range.is_empty());
assert_ne!(range.start, range.end);

range.next();

assert!(range.is_empty());
assert_eq!(range.start, range.end);

pub fn contains<M2, T2>(&self, pointer: &BitPtr<M2, T2, O>) -> bool
where M2: Mutability, T2: BitStore,

Tests if a given bit-pointer is contained within the range.

Bit-pointer ordering is defined when the types have the same exact BitOrder type parameter and the same BitStore::Mem associated type (but are free to differ in alias condition!). Inclusion in a range occurs when the bit-pointer is not strictly less than the range start, and is strictly less than the range end.

Original

Range::contains

Examples

use bitvec::prelude::*;
use bitvec::ptr::BitPtrRange;
use core::cell::Cell;

let data = 0u16;
let bp = BitPtr::<_, _, Lsb0>::from_ref(&data);

let mut range = BitPtrRange::from_range(bp .. bp.wrapping_add(16));
range.nth(2);
range.nth_back(2);

assert!(bp < range.start);
assert!(!range.contains(&bp));

let mid = bp.wrapping_add(8);

let same_mem = mid.cast::<Cell<u16>>();
assert!(range.contains(&mid));

Casting to a different BitStore type whose Mem parameter differs from the range always results in a false response, even if the pointer being tested is numerically within the range.