Struct common_base::bit_vec::prelude::BitRef

#[repr(C, align(8))]
pub struct BitRef<'a, M = Const, T = usize, O = Lsb0>
where
    M: Mutability,
    T: BitStore,
    O: BitOrder,
{
    bitptr: BitPtr<M, T, O>,
    data: bool,
    _ref: PhantomData<&'a UnsafeCell<bool>>,
}

Proxy Bit-Reference

This structure simulates &/mut bool within BitSlice regions. It is analogous to the C++ type std::bitset<N>::reference.

This type wraps a BitPtr and caches a bool in one of the remaining padding bytes. It is then able to freely give out references to its cached bool, and commits the cached value back to the proxied location when dropped.

Original

This is semantically equivalent to &'a bool or &'a mut bool.

Quirks

Because this type has both a lifetime and a destructor, it can introduce an uncommon syntax error condition in Rust. When an expression that produces this type is in the final expression of a block, including if that expression is used as a condition in a match, if let, or if, then the compiler will attempt to extend the drop scope of this type to the outside of the block. This causes a lifetime mismatch error if the source region from which this proxy is produced begins its lifetime inside the block.

If you get a compiler error that this type causes something to be dropped while borrowed, you can end the borrow by putting any expression-ending syntax element after the offending expression that produces this type, including a semicolon or an item definition.

Examples

use bitvec::prelude::*;

let bits = bits![mut 0; 2];

let (left, right) = bits.split_at_mut(1);
let mut first = left.get_mut(0).unwrap();
let second = right.get_mut(0).unwrap();

// Writing through a dereference requires a `mut` binding.
*first = true;
// Writing through the explicit method call does not.
second.commit(true);

drop(first); // Itโs not a reference, so NLL does not apply!
assert_eq!(bits, bits![1; 2]);

Fields
bitptr: BitPtr<M, T, O>data: bool_ref: PhantomData<&'a UnsafeCell<bool>>
Implementations

impl<M, T, O> BitRef<'_, M, T, O>
where M: Mutability, T: BitStore, O: BitOrder,

pub unsafe fn from_bitptr(bitptr: BitPtr<M, T, O>) -> BitRef<'_, M, T, O>

Converts a bit-pointer into a proxy bit-reference.

This reads through the pointer in order to cache the current bit value in the proxy.

Original

The syntax unsafe { &* ptr }.

Safety

This is equivalent to (and is!) dereferencing a raw pointer. The pointer must be well-constructed, refer to a live memory location in the program context, and not be aliased beyond its typing indicators.

pub fn into_bitptr(self) -> BitPtr<M, T, O>

Decays the bit-reference to an ordinary bit-pointer.

Original

The syntax &val as *T.

impl<T, O> BitRef<'_, Mut, T, O>
where T: BitStore, O: BitOrder,

pub fn replace(&mut self, src: bool) -> bool

Moves src into the referenced bit, returning the previous value.

Original

mem::replace

pub fn swap<T2, O2>(&mut self, other: &mut BitRef<'_, Mut, T2, O2>)
where T2: BitStore, O2: BitOrder,

Swaps the bit values of two proxies.

Original

mem::swap

pub fn commit(self, value: bool)

Commits a bit into the proxied location.

This function writes value directly into the proxied location, bypassing the cache and destroying the proxy. This eliminates the second write done in the destructor, and allows code to be slightly faster.

pub fn set(&mut self, value: bool)

Writes value into the proxy.

This does not write into the proxied location; that is deferred until the proxy destructor runs.