Trait common_base::buffer::Buffer

pub trait Buffer {

    // Required methods
    fn remaining_size(&self) -> usize;
    fn peek_to_slice(&self, dst: &mut [u8]) -> Result<()>;
    fn advance_by(&mut self, by: usize);

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn read_to_slice(&mut self, dst: &mut [u8]) -> Result<()> { ... }
    fn read_u8_le(&mut self) -> Result<u8> { ... }
    fn peek_u8_le(&mut self) -> Result<u8> { ... }
    fn read_i8_le(&mut self) -> Result<i8> { ... }
    fn peek_i8_le(&mut self) -> Result<i8> { ... }
    fn read_u16_le(&mut self) -> Result<u16> { ... }
    fn peek_u16_le(&mut self) -> Result<u16> { ... }
    fn read_i16_le(&mut self) -> Result<i16> { ... }
    fn peek_i16_le(&mut self) -> Result<i16> { ... }
    fn read_u32_le(&mut self) -> Result<u32> { ... }
    fn peek_u32_le(&mut self) -> Result<u32> { ... }
    fn read_i32_le(&mut self) -> Result<i32> { ... }
    fn peek_i32_le(&mut self) -> Result<i32> { ... }
    fn read_u64_le(&mut self) -> Result<u64> { ... }
    fn peek_u64_le(&mut self) -> Result<u64> { ... }
    fn read_i64_le(&mut self) -> Result<i64> { ... }
    fn peek_i64_le(&mut self) -> Result<i64> { ... }
    fn read_f32_le(&mut self) -> Result<f32> { ... }
    fn peek_f32_le(&mut self) -> Result<f32> { ... }
    fn read_f64_le(&mut self) -> Result<f64> { ... }
    fn peek_f64_le(&mut self) -> Result<f64> { ... }

}

Required Methods

fn remaining_size(&self) -> usize

Returns remaining data size for read.

fn peek_to_slice(&self, dst: &mut [u8]) -> Result<()>

Peeks data into dst. This method should not change internal cursor, invoke advance_by if needed.

Panics

This method may panic if buffer does not have enough data to be copied to dst.

fn advance_by(&mut self, by: usize)

Advances internal cursor for next read.

Panics

This method may panic if the offset after advancing exceeds the length of underlying buffer.

Provided Methods

fn is_empty(&self) -> bool

Returns true if buffer has no data for read.

fn read_to_slice(&mut self, dst: &mut [u8]) -> Result<()>

Reads data into dst. This method will change internal cursor.

Panics

This method may panic if buffer does not have enough data to be copied to dst.

fn read_u8_le(&mut self) -> Result<u8>

fn peek_u8_le(&mut self) -> Result<u8>

fn read_i8_le(&mut self) -> Result<i8>

fn peek_i8_le(&mut self) -> Result<i8>

fn read_u16_le(&mut self) -> Result<u16>

fn peek_u16_le(&mut self) -> Result<u16>

fn read_i16_le(&mut self) -> Result<i16>

fn peek_i16_le(&mut self) -> Result<i16>

fn read_u32_le(&mut self) -> Result<u32>

fn peek_u32_le(&mut self) -> Result<u32>

fn read_i32_le(&mut self) -> Result<i32>

fn peek_i32_le(&mut self) -> Result<i32>

fn read_u64_le(&mut self) -> Result<u64>

fn peek_u64_le(&mut self) -> Result<u64>

fn read_i64_le(&mut self) -> Result<i64>

fn peek_i64_le(&mut self) -> Result<i64>

fn read_f32_le(&mut self) -> Result<f32>

fn peek_f32_le(&mut self) -> Result<f32>

fn read_f64_le(&mut self) -> Result<f64>

fn peek_f64_le(&mut self) -> Result<f64>