common_base::hash

Struct FixedRandomState

pub struct FixedRandomState {
    state: RandomState,
}

A random state with fixed seeds.

This is used to ensure that the hash values are consistent across different processes, and easy to serialize and deserialize.

Fields
state: RandomState
Implementations

impl FixedRandomState

const RANDOM_SEED_0: u64 = 5_871_781_006_564_002_453u64

const RANDOM_SEED_1: u64 = 4_794_697_086_780_616_226u64

const RANDOM_SEED_2: u64 = 8_158_064_640_168_781_261u64

const RANDOM_SEED_3: u64 = 13_096_744_586_834_688_815u64

pub fn new() -> Self