common_base::range_read

Struct AsyncReadAdapter

pub struct AsyncReadAdapter<R> {
    inner: Arc<Mutex<R>>,
    position: u64,
    buffer: Vec<u8>,
    content_length: u64,
    read_fut: Option<Pin<Box<dyn Future<Output = Result<Bytes>> + Send>>>,
}

AsyncReadAdapter adapts a RangeReader to an AsyncRead.

Fields
inner: Arc<Mutex<R>>

The inner RangeReader. Use Mutex to get rid of the borrow checker issue.

position: u64

The current position from the view of the reader.

buffer: Vec<u8>

The buffer for the read bytes.

content_length: u64

The length of the content.

read_fut: Option<Pin<Box<dyn Future<Output = Result<Bytes>> + Send>>>

The future for reading the next bytes.

Implementations

impl<R> AsyncReadAdapter<R>

pub(crate) fn project<'pin>( self: Pin<&'pin mut Self>, ) -> __AsyncReadAdapterProjection<'pin, R>

pub(crate) fn project_ref<'pin>( self: Pin<&'pin Self>, ) -> __AsyncReadAdapterProjectionRef<'pin, R>

impl<R: RangeReader + 'static> AsyncReadAdapter<R>

pub async fn new(inner: R) -> Result<Self>